Improve kserve protocol version handling (#2957)
* fix(kserve): ensure there's a default protocol configured

The current implementation retrieves the protocol to use from the
"PROTOCOL_VERSION" environment variable however there's no default value
which will trigger an error when served through Kserve as the base class
does a protocol check in the predict method that will fail with None.

The default protocol uses the same value as the base class.

* fix(kserve): ensure the protocol version configured is a valid value

* feat(kserve): make configuration file path configurable

This will allow to make the wrapper easier to test beside making
it possible to to change where the file should be looked for.

* test: add KServe wrapper test

* test(kserve_wrapper): add protobuf code generation

* fix(kserse): add None handling to the wrapper

In case None is passed, keep the default value set in the base __init__.

This makes TorchserveModel behave in the same fashion as the base class.

* refactor(test): rewrite wrapper test to be more complete

It now validates that the protocol version is passed properly as well
as failure if an invalid protocol is given.

* test: remove kserve pytest tests

The overhead required to run the tests outweighs it benefits.

The thing to keep in mind is that the fix allows to run models
through kserve deployments prior to 0.11.1.